% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bins.R
\name{bins_ss}
\alias{bins_ss}
\title{Calculate degree bins using the single sine method.}
\usage{
bins_ss(t0, t1, tmin, tmax, weights = NULL, parallel = FALSE)
}
\value{
\code{data.frame} with degree days.
}
\description{
C++ function for calculating a data.frame
of degree bins for vectors of tmin / tmax, and vectors of bins bounds.
}

