% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwd.R
\name{calc_monthly_mean}
\alias{calc_monthly_mean}
\title{Calculate Monthly Mean}
\usage{
calc_monthly_mean(df_hdd, from = 2011, to = 2020)
}
\arguments{
\item{df_hdd}{dataframe from function get_montly_hdd; should be tidyed before}

\item{from}{first year}

\item{to}{last year}
}
\description{
Calculate monthly mean for given period
}
