% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{days_in_bin_daily}
\alias{days_in_bin_daily}
\title{Calculate days in bin for daily data.}
\usage{
days_in_bin_daily(t0, t1, tmin, tmax)
}
\arguments{
\item{t0}{vector of lower bounds}

\item{t1}{vector of upper bounds}

\item{tmin}{vector of tmin values (1 per day)}

\item{tmax}{vector of tmax values (1 per day)}
}
\value{
num_days x num_bins \code{matrix}
}
\description{
Calculate days in bin for daily data.
}

