% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dd.R
\name{dd_band_ss}
\alias{dd_band_ss}
\title{Calculate degree days using the single sine method.}
\usage{
dd_band_ss(t0, t1, tmin, tmax, parallel = FALSE)
}
\arguments{
\item{t0}{vector of lower bounds}

\item{t1}{vector of upper bounds}

\item{tmin}{vector of tmin values (1 per day)}

\item{tmax}{vector of tmax values (1 per day)}
}
\value{
\code{data.frame} with degree days.
}
\description{
C++ function for calculating a data.frame
of degree days for vectors of tmin / tmax, and vectors of degree day bands.
}

