% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{degree_days_band_daily}
\alias{degree_days_band_daily}
\title{Calculate degree days for daily data.}
\usage{
degree_days_band_daily(t0, t1, tmin, tmax)
}
\arguments{
\item{t0}{vector of lower bounds}

\item{t1}{vector of upper bounds}

\item{tmin}{vector of tmin values (1 per day)}

\item{tmax}{vector of tmax values (1 per day)}
}
\value{
num_days x num_bins \code{matrix}
}
\description{
Calculate degree days for daily data using C++ code.
}

