% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwd.R
\name{download_dwd_data}
\alias{download_dwd_data}
\title{get monthly hdd}
\usage{
download_dwd_data()
}
\description{
DWD Climate Data Center (CDC): Recent monthly degree days according
to VDI 3807 for Germany, quality control not completed yet, version v19.3
}
