% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwd.R
\name{download_dwd_data_single_file}
\alias{download_dwd_data_single_file}
\title{Download a single month from DWD server}
\usage{
download_dwd_data_single_file(url)
}
\arguments{
\item{url}{url to the file on dwd server}
}
\description{
Download a single month from DWD server
}
