% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polys.R
\name{poly_ss}
\alias{poly_ss}
\title{Calculate degree polynomials using the single sine method.}
\usage{
poly_ss(tmin, tmax, degree, weights = NULL, parallel = FALSE)
}
\arguments{
\item{tmin}{vector of tmin values (1 per day)}

\item{tmax}{vector of tmax values (1 per day)}

\item{weights}{vector of optional weights to multiply output by. Default is 1.}
}
\value{
\code{data.frame} with degree days.
}
\description{
C++ function for calculating a data.frame
of degree polynomials for vectors of tmin / tmax.
}

