% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sin_estimate_NumericVector}
\alias{sin_estimate_NumericVector}
\title{Sine function with max and min given.}
\usage{
sin_estimate_NumericVector(x, tmin, tmax)
}
\arguments{
\item{x}{Vector of x values in [0, 1] to evaluate at.}

\item{tmin}{\code{double} tmin value.}

\item{tmax}{\code{double} tmax value.}
}
\value{
Vector of interpolated sine values.
}
\description{
Sine function evaluated at x for single sine interpolation.
}
