% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{spl1_band_daily}
\alias{spl1_band_daily}
\title{Calculate degree days/degree-1 splines for daily data.}
\usage{
spl1_band_daily(t0, t1, tmin, tmax, weights)
}
\arguments{
\item{t0}{vector of lower bounds}

\item{t1}{vector of upper bounds}

\item{tmin}{vector of tmin values (1 per day)}

\item{tmax}{vector of tmax values (1 per day)}

\item{weights}{vector of optional weights to multiply output by. Default is 1.}

\item{tmax}{vector of tmax values (1 per day)}
}
\value{
num_days x num_bins \code{matrix}
}
\description{
Calculate degree days/degree-1 splines for daily data using C++ code.
}

