% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd.R
\name{spl1_band_ss}
\alias{spl1_band_ss}
\title{Calculate degree days/degree-1 splines of temperature using the single sine method.}
\usage{
spl1_band_ss(t0, t1, tmin, tmax, weights = NULL, parallel = FALSE)
}
\arguments{
\item{t0}{vector of lower bounds}

\item{t1}{vector of upper bounds}

\item{tmin}{vector of tmin values (1 per day)}

\item{tmax}{vector of tmax values (1 per day)}

\item{weights}{vector of optional weights to multiply output by. Default is 1.}
}
\value{
\code{data.frame} with degree days.
}
\description{
C++ function for calculating a data.frame
of degree days/degree-1 splines for vectors of tmin / tmax.
}
