% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{spl3_band_daily}
\alias{spl3_band_daily}
\title{Calculate degree days/degree-3 splines for daily data.}
\usage{
spl3_band_daily(t0, tmin, tmax, weights)
}
\arguments{
\item{t0}{vector of lower bounds}

\item{tmin}{vector of tmin values (1 per day)}

\item{tmax}{vector of tmax values (1 per day)}

\item{weights}{vector of optional weights to multiply output by. Default is 1.}
}
\value{
num_days x num_bands \code{matrix}
}
\description{
Calculate degree days/degree-3 splines for daily data using C++ code.
}
