% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwd.R
\name{tidy_hdd}
\alias{tidy_hdd}
\title{tidy hdd}
\usage{
tidy_hdd(df_hdd, threshold = 0.95)
}
\arguments{
\item{df_hdd}{dataframe from function get_montly_hdd}

\item{threshold}{keep only stations with sufficient data (completeness ≥ threshold)}
}
\description{
Some opinionated tidying of the heating degree days dataframe:
\itemize{
\item convert to tsibble
\item keep only stations with sufficient data (completeness ≥ threshold)
}
}
