% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bins.R
\name{bins_ss}
\alias{bins_ss}
\title{Calculate degree bins using the single sine method.}
\usage{
bins_ss(t0, t1, tmin, tmax, weights = NULL, parallel = FALSE)
}
\arguments{
\item{t0}{vector of lower bounds}

\item{t1}{vector of upper bounds}

\item{tmin}{vector of tmin values (1 per day)}

\item{tmax}{vector of tmax values (1 per day)}

\item{weights}{vector of optional weights to multiply output by. Default is 1.}
}
\value{
\code{data.frame} with degree days.
}
\description{
C++ function for calculating a data.frame
of degree bins for vectors of tmin / tmax, and vectors of bins bounds.
}

